Apply scopes from diagnostics t references output panel highlighting
The references output is quite static so multiple pushes may make things too complicated here.
The scope names are applied the same way we did for diagnostics.

The tab between file path and location is changed into a normal whitespace as tab does not align correctly if the file paths' length in the list differs too much. It looks better with fixed distance.

We may or may not scope the preamble more detailed, but I find a more minimalistic scoping better at the moment.