Store output panel settings in a dictionary and apply in a loop
The settings and values to be applied to each output panel are stored in a dictionary, which may or may not be outsourced to a resource file later.

Most of the settings are "required" constants to make the output panel look as it does. Therefore they are not intended to be modified by user. Some other settings such as row padding, margin, or maybe some other may be changed by users choice. But only few will probably care about.

The "result_base_dir" setting may change during the lifetime of the panel and therefore needs to be updated everytime the content is updated. A change may happen, if a new project is opened in a window.

TODO: The output panel may show the results of several views, which may show files from different folders. Therefore a global "result_base_dir" may not work for all displayed files.

This means the output panel can either display the content of one view only or we need to handle file navigation events on our own with a dedicated EventListener.