Create LspSettings class to manage cached settings
With respect to object oriented programming which is already used for most of the language-server-protocol based stuff, the configuration of the LSP package should follow this approach, if it is intended to cache the settings.

Therefore a class LspSettings is created, which uses a template LSP_SETTINGS to initialize the object attributes at startup. In plugin_loaded() callback the real values from the LSP.sublime-settings are loaded, mapped to the object attributes and kept up to date by the subscribed the on_change event.

The one and only place to define settings to handle is the LSP_SETTINGS dictionary.

All managed settings can be accessed (read-only) as normal attribute by calling

   lsp_settings.my_setting_from_lsp_sublime_settings

The only drawback is mypy which complains about class attributes not found. Therefore several '# type: ignore' statements need to be added.