Always show symbol information popup (#46)
* Don't fight with other visible popups and display hint if nothing was found.

on_hover() should not create a tooltip, if one is visible already to avoid popup fighing and flickering.

If the server response object or the contents returned is empty, display a tooltip with "No description available." message.

* Add symmetrical spacing around paragraphs in popups.

mdpopups adds extra spacing to the bottom of a paragraph. This may be a good thing for longer texts but makes single lined popups look a bit sluggish. Therefore we should add a symmetrical spacing around it to still keep spacing between paragraphs.

* Exclude some scopes from symbol information query.

Some scopes will most likely point to code entities which a languages server won't provide information for. Those scopes can safely be used to prevent unnecessary requests being sent. Maybe some per language server setting may be used in the future to adjust the filters.