Fix typing errors and raise exceptions for missing crucial response entries
This commit ensures correct typing of arguments when creating Point and Range objects to satisfy mypy.

The main issue with the previous commit was the usage of the method `.get()`, which may be used for optional parts of a JSON-RPC response only. The mypy error is resolved by accessing the crucial point/range properties with `point['line']`, ... .

These changes are made to all appropriate parts of the code where range/point objects of a response message are expected in order to ensure unpredicted behavior due to invalid range/point objects.

As an result a malformed response message from a language-server will correctly raise

		- `TypeError` if a range/point object is missing,
		- `KeyError` if a required child object of a range/point is missing
		- `ValueError` if `"line"` or `"character"` are not of type integer

These exceptions are displayed in the console as:

	"LSP: Error handling server content: ..."

The message may be made a bit more verbose in the future.