Exclude some stuff from sublime-package files
Some of the files and folders of the repository are meant for development only and don't need to be part of a sublime-package file. We may therefore add an .gitattributes file to exclude them from being archived/exported. This may reduce the package size somewhat and makes the package content look clean.