Do some type runtime checking by casting the values to int.
mypy does static type checking only, but can't prevent malformed responses to be passed to LSP nor check types during runtime.

Therefore some checks may help to reject malformed responses and prevent unpredictable behavior of the LSP package, which might in worst case lead to unwanted text manipulations or something like that, we'd to be blamed for in the future.

An external server/application should not be considered trustworthy.