Add textDocument/willSave and textDocument/willSaveWaitUntil
For willSaveWaitUntil, it is required to *wait* in the on_pre_save ST callback
in order to synchronize formatting&saving. For this reason the functionality
of a request has been enhanced with a blocking mode (by default set to False)
that blocks the requesting thread until the server responds. A timeout of six
seconds is in place in case the server doesn't respond.