Add a setting to display either completion kind or details
Prior to this commit the completion panel always displays the text provided
by the "details" item of the language server response and falls back to
the item kind only if "details" is not available.

Some language-servers provide quite long details which can't be displayed
correctly by ST's completion panel and in some cases don't even provide
further information as they are more or less a copy of `label`.

Therefore this commit adds a setting `show_completion_details` which is set
`true` by default to enable LSP to always display the completion kind rather
than details.

To ensure, to gracefully fail, if none of the fields is available, the
CompletionItemKind object is extended to provide an `Unknown` item.

The label and the details of a completion item returned by `format_completion()`
are separated by ` \t ` rather than `\t` as ST tends to display no space
between them if the details string is too long otherwise.