Add support for PackageDev settings tooltips
PackageDev uses the comments above each setting to display a tooltip,
if the cursor is hovered over an existing setting.

This commit adds required comments to the LSP.sublime-settings file.

An empty "clients" object was added to prevent PackageDev's linter from
highlighting the key as unknown.

The template for user defined clients is moved to the "clients" setting
and a warning was added to the comment of "default_clients" to prevent
a user from accidentally manipulating the wrong setting to do overrides.

The template was formated to hopefully display well in the tooltip for
most users.