Fix word completions not displayed in completions panel
fixes #158

The format of the "trigger" of the completions returned by a plugin
seems to influence the way Sublime Text merges them with word completions.

The extra spaces added to the "trigger" value, which make sure the hints
are displayed with enough distance to the label, cause Sublime Text
not to correctly merge in word completions.

To solve the issue, the `\t` is moved to the very left to prevent
a space to be appended to the label in the returned value.