Handle 'null' result in JSON RPC
A few methods in the Language Server Protocol can return `null`.
The Python library maps the JSON value `null` to `None`. So far,
rpc.py was unable to distinguish between `{result: null}` and a missing
result field. It reported both of them as protocol violation.

Due to this, it was not possible to restart the python-language-server
cleanly. Trying to execute `Restart Servers` always resulted in an
`incomplete response payload` error since the `shutdown` request always
returns `null`.

This commit adapts rpc.py so that it accepts `{result: null}`. In case
both `result` and a non-empty `error` are received, we treat that as an
error response.

Also, a response of the form of `{result: null, error: {code: 123,
message: "error"}}` is treated as invalid. According to the JSON-RPC spec,
it is invalid.