Fix: _is_visible property of SignatureHelpListener is unreliable
Replace it with self.view.is_popup_visible. The downside is that this method
also return `True` when the signature help popup is hidden, but the on-hover
popup is visible. Thus, we try to make sure that the signature help popup is
"truly" visible/hidden when the _signatures list is empty.