Send what features we support to the server. (#279)
* Send what features we support to the server.
Some servers will listen to what features a client says it supports and will disable those features, returning "False" in the server capabilities reply.  Which, in turn, disables that feature in our code.

* The URI conversion will sometimes fail on file:// handlers.
Sometimes, a Windows pathname will be converted incorrectly.  This is a bandaid fix on the problem.

Previously, we would a path that started like this: \C:\Users\...

When we tossed that path into Sublime, it would not be able to find the view associated to that file.  This bandaid fix would remove the preceding \ in that pathname.

This will mainly fix issues with diagnostics.

* Pass build tests.

* Use the class objects instead of magic numbers.

* Send parent process id in initialization params.
According to the LSP specifications, the processId field in the initialization params needs to be the parent process id.  This is so the language server can shut itself down with the parent process terminates.  This will fix bugs where when you close Sublime Text, the language server would be kept running.

* Pass build tests.

* Indicate that we support markdown on signatures.

* Support markdown for hover.

* Revert "Merge branch 'master' of github.com:LoneBoco/LSP into send-supported-features"

This reverts commit a3a1d8f0e749703f2bfbab2be3365e8d654c73bd, reversing
changes made to 2ed9bd3336600ffe040ea149b46cf2b15dea2895.