Fire request on selection modified, on TextCommand and after Did Change event
Request needs to be send on selection modified, to determine whether to show the bulb or not by the caret position.

The request needs to be send when the TextCommand is run, because the command might be triggered from the phantom (error|warning|hint) pop-up and when it is triggered the caret might already be on a place where code action exist, so the mouse will trigger those code actions instead of the ones that were suppose to be triggered from the popup.

And finally the request needs to be send after the `textDocument/didChange`. So the code actions show suggestions for the current version of document and not for an outdated one.