Introduce cache
Code actions get cached on each request, so when running the TextCommand the menu will be populated my cache.

However, hovers for warning, errors, hint can run the code action text command. In that case we need to fire a new request, by passing a flag `make_request`, so the code actions that will be displayed are related to the point from where the hover was triggered.