Add a bulb (#489)
* Add a bulb

* Delete unneeded import and switch from absolute to relative import

* Fire request on selection modified, on TextCommand and after Did Change event

Request needs to be send on selection modified, to determine whether to show the bulb or not by the caret position.

The request needs to be send when the TextCommand is run, because the command might be triggered from the phantom (error|warning|hint) pop-up and when it is triggered the caret might already be on a place where code action exist, so the mouse will trigger those code actions instead of the ones that were suppose to be triggered from the popup.

And finally the request needs to be send after the `textDocument/didChange`. So the code actions show suggestions for the current version of document and not for an outdated one.

* Introduce cache

Code actions get cached on each request, so when running the TextCommand the menu will be populated my cache.

However, hovers for warning, errors, hint can run the code action text command. In that case we need to fire a new request, by passing a flag `make_request`, so the code actions that will be displayed are related to the point from where the hover was triggered.

* Fix formating

* add icons

* format import

* Don't show the command in the right click menu, when 0 code actions

This is different from is_enabled because I don't want to disable the command from running, so the user get the `no code actions` popup

* Fix formating

* Remove return type

* Ignore type, https://github.com/python/mypy/issues/500#issuecomment-87460993

Don't know how to typehint debounce.t

* return None for __init__

* leave hover unchanged

* Remove cache and event listener, Add a command to update code actions

* Need to be at least 500ms

To avoid messages in status bar like `Request textDocument/codeAction failed with message. Debug failure. Expected 80 <= 50`

* Don't fire command in windows.py file

Instead, listen in code_actions.py on_modified_async.

An explanation.
The following will explain why we need to listen on changes?

Short Explanation
So we don't end up with a bulb in the gutter when there are no code actions.

Long (Detailed) explanation
1. When on_modified_async is fired, on_selection_modified_async is fired as well.

2. When inserting text,
I need to make sure that textDocument/codeAction is fired after textDocument/didChange

Because of this line in windows.py, notice the 0.5s delay.
```python
self._sublime.set_timeout_async(lambda: self.purge_did_change(buffer_id, buffer_version), 500)
```
source: https://github.com/tomv564/LSP/blob/41eba6ef3ca1db9fb2ab4f3b393afe4e376b4c40/plugin/core/windows.py#L256

That is the reason the handle_modified_async fire_requsest has a debounce of 0.8s,
0.3s after from textDocument/didChange.

3. If handle_modified_async is fired,
it will debounce fire_request call for another 0.3s (0.8s of handle_modified_async - 0.5s of fire_request)

It is needed to check if the code actions are valid after we made some changes.
So we don't end up with a bulb in the gutter when there are no code actions

* Add a flag

* Leave windows.py unchanged

* Rename show_bulb to show_code_actions_bulb

* Hide bulb when  on_selection_modified_async is fired and remove on_modified_async listener.

For that change I needed to increase the debounce for fire_request from 0.5 to 0.8 ().

Removed LspUpdateCodeActionsCommand, no need for it

Pros:
Simpler code, with not so noticeable changes for the user
Cons:
Fires 300ms slower for on_selection_modified_async, that it used to do, but not so noticeable as said

* Reintroduce on_modified_async, because of a better user experience

* Revert "Reintroduce on_modified_async, because of a better user experience"

This reverts commit ccef8b7337bf042a02cefbfa569c1aa1417437a0.