Test text edit ranges that don't overlap but touch
The spec for `TextEdit[]` reads: "All text edits ranges refer to
positions in the original document. Text edits ranges must never
overlap, that means no part of the original document must be manipulated
by more than one edit. However, it is possible that multiple edits have
the same start position: multiple inserts, or any number of inserts
followed by a single remove or replace edit. If multiple inserts have
the same position, the order in the array defines the order in which the
inserted strings appear in the resulting text."