Introduce "completion_trigger" setting.
The completion_trigger setting decides what the trigger of the completion will
be. It can be one of:

1) label,
2) filterText,
3) the replacement text.

By default, it is set to the replacement text. This results in a noisy
auto-complete widget. However, it makes most of the known language servers work
correctly, despite ST not supporting TextEdit.