Teach LSP about workspaces
Introduce the Workspace class in protocol.py. It has a name and a URI.
Refactor most of plugin/core/workspace.py to work with the new Workspace
class.

In the WindowManager class, use the new Workspace class at the boundary
of the session starter callable.

So, the Session class doesn't yet understand what a workspace is, but
the WindowManager does.

This diff does absolutely nothing feature-wise, it just refactors code.