Teach LSP about WorkspaceFolder (#770)
* Teach LSP about workspaces

Introduce the Workspace class in protocol.py. It has a name and a URI.
Refactor most of plugin/core/workspace.py to work with the new Workspace
class.

In the WindowManager class, use the new Workspace class at the boundary
of the session starter callable.

So, the Session class doesn't yet understand what a workspace is, but
the WindowManager does.

This diff does absolutely nothing feature-wise, it just refactors code.

* Remove get_common_prefix, store path instead of URI

* Rename Workspace to WorkspaceFolder

* Use .folders() instead of .project_data() to determine path and name for WorkspaceFolder

* Fix mypy errors hopefully

* Fix pytest tests

* Allow a Window to be passed to maybe_get_workspace_from_view

The maybe_get_workspace_from_view function will check if it's dealing
with a window or a view. If it's dealing with a window, it will use
the active_view() of that window.

* 'type: ignore' is not needed

* Remove unused import