Prevent a deadlock by running all handlers from the ST3 alternate thread
When quickly editing and saving, it is possible that e.g. a completion
request is sent and then a formatting request.

When that happens, the handler for the completion request is invoked
while we're blocking the main thread. But in the handler we bring up
the completion widget. So that means the handler will block. So the
reader thread blocks. Then a timeout occurs.

To prevent this from happening, schedule all handlers on the ST3 alternate
thread. The RPC reader thread is unblocked and is ready to process more
notifications/responses. The response for the blocking request should
then come next and the main thread is unblocked properly.

- Improve SyncRequestStatus class; it doesn't need a __state property