Server logs to panel (#811)
* Implement a general-purpose panel toggler commmand

* Replace 'lsp_show_diagnostics_panel' with 'lsp_toggle_panel'

* Working towards logging to a panel

It only displays the last line now.
It should become a circular buffer.
Look into using sublime_lib.

* Add sublime_lib dependency

* First steps with sublime_lib

Use the Panel class to toggle visibility

* Use an OutputPanel from sublime_lib to print server messages to the panel

It apparently has an infinite buffer so we should make that circular instead.
You can click inside the panel and that screws up the output.
Should not make it interactive.

* Force writes, seek cursor to end before writing logs

No need for set_read_only anymore

* Convert the panel handling in references.py to using an OutputPanel

* Convert te panel handling in diagnostics.py to using a sublime_lib.OutputPanel

* Remove unused import

* All our panels are read-only

* Don't forget to destroy the server log panel when the plugin is disabled

* Remove LspClearPanelCommand

It is no longer used.

* Remove LspUpdatePanelCommand

It is no longer used.

* Remove the server_log function

It is unused.

* Fix error invoking log_callback during merge

* Rework lsp_toggle_panel command

* Revert "Convert te panel handling in diagnostics.py to using a sublime_lib.OutputPanel"

This reverts commit 558f48918ac636393eb770d9c4fb317cd6761ab2.

* Revert "Convert the panel handling in references.py to using an OutputPanel"

This reverts commit d539380bbf864260bd6cd4fcc780ace62209a577.

* Revert "Remove LspUpdatePanelCommand"

This reverts commit c8867c43ed328a8e6ed825e8663d7f93a08e47bf.

* Revert "Remove LspClearPanelCommand"

This reverts commit 7b83acfbf4018efa6c5273184880b3c7ebdeeabe.

* Replace OutputPanel with lsp_update_server_panel command

The command modifies the panel atomically, preventing jankyness

* Print LSP payloads to the server panel

* Align notifications with the other payload types ('<-- ' vs. '<--')

* Syntax highlight the server logs panel

* Do not print the payloads of textDocument/did{Change,Open}

They are too large and ST will refuse to syntax-highlight long lines anyway.

* Revert "First steps with sublime_lib"

This reverts commit 22112323b42d89869a783ce0233f7abdda857c9b.

* Fix mypy errors

* Revert dependencies.json

* Add a test for the circular behavior of the server log panel

* Enable word-wrap for server logs

* Install the stderr log callback through the "session starter" callable

* Do not enable "word_wrap" for now

I get a few hickups in the UI when "word_wrap" is enabled in the panel.
I'm not sure how to deal with this.

* Fine tuning

* Use MockSettings in the pytest tests

* Wrong import location for ensure_server_panel in the test

* Oops, accidentally disabled all debug logging. It is restored now

* Also print unhandled notifications/requests to the server panel

* Update LSP.sublime-settings

* Fix mock Client class

It was missing a method, causing an AttributeError in the session starter

* Fix sync requests not working for the server panel

Also, only hook up the log payload handler if we actually have a panel
available to us (a.k.a. don't hook it up in tests)

* Remove #noqa line, used to be too long but not anymore

* Fix broken test depending on external state

Run "Test Current File" from the command palette twice to see the effect.

* Deprecate lsp_show_diagnostics_panel command

Print a message to the console for users that have this command mapped
to some other keybinding in their user prefs.

* Rename panel to "Language Servers" instead of "Server"

* Less monkey-typing, more functions for test_server_panel_circular.py

* Put the name of the server after the arrow

Replace the word "LSP" with just a colon

* Remove the intro message, it seems a bit redundant

* Also account for payload results that are lists, and payload results that are null

* Lookahead for any non-whitespace character instead of a '{'

in order to match "None" or the beginning of a list '['

* Cleanly destroy panels

Before this change, you'd get an error popup about a missing syntax file.
So, change the syntax to plain text just before destroying the output panel.

This ensures a clean shutdown of the package (w.r.t. panels at least)

* Put "lsp_active": True in all panels

So that keybindings with a context keep working if the user's cursor
is in the panel somewhere.

* Don't modify settings.py

* Minimally modify boot.py

* Compactify imports

* Remove debug printing trying to figure out how panels are unloaded

* Don't make the WindowManager do all the work that the Client already did

Instead, we have a logger class now with methods such as
outgoing_request, incoming_request.

The WindowManager modifies the eventual log sink of that object so
that it'll log to a panel.

Interestingly, this also makes it easy to log to the console again.

* Just remove the Payload type entirely and replace with Any

* Just write two separate commands for the panel togglers

* Fix overindentation

* Ensure an int for request_id in rpc.py, it might be a string!

* struck by flakiness yet again