Read values for workspace/configuration request from settings
See: https://microsoft.github.io/language-server-protocol/specification#workspace_configuration

The Julia server sends multiple configuration keys as 'section'.
This commit extracts the corresponding values for those keys from the
"settings" entry of the server configuration, and returns None for
missing values, as described in the protocol specification.

A special case is needed for the ESLint server, which sends a blank
section instead, but expects a dict of configuration settings.