Workaround for TextEdit range not being valid when selection a completion item
The workaround works like this.

First a fact: "The specs says that a TextEdit must span only on one line".

when we request completion items.
and get a response.

We save each line where a cursor is.
When we select a command completion item.
The text is deleted.

But becuse we saved the line contents before, we just restore it. so we have the state as before.

This workaround have one big downside, I guess it wouldn't work performant for lines that are really really really long.