Translate in UTF-16 points
The functions point_to_offset offset_to_point also need to be translated
in terms of UTF-16 column offsets. Unfortunately, I couldn't find a way
to clamp the column anymore. Using sublime.View.line gives us the
number of UTF-32 points, not the number of UTF-16 points (which is
what we need). I've added a TODO comment for this.