Maintenance/merge master into st4000-exploration (#938)
* Restart the server when it fails on exit

* Add 'actions' class to hover popup action links.

The popups.css selector uses ".lsp_popups .actions a" which does not
match the generated html. Without this, the links are solid blue without
framing.

* Add support for Swift's sourcekit-lsp

* Use line-wrapping algorithm provided by mdpopups (#925)

* Use line-wrapping algorithm provided by mdpopups

It seems to only break lines if it finds opportunity to
do so (whitespace). That means that in some cases lines won't
be wrapped where old code would aggressively wrap.

Resolves #859

* Narrow down imports (#919)

* Update message for 3154.0.0

* Fix issue after merge

Co-authored-by: Paul <pfultz2@yahoo.com>
Co-authored-by: Francois H. Theron <16208294+FHTheron@users.noreply.github.com>
Co-authored-by: Saagar Jha <saagar@saagarjha.com>
Co-authored-by: Raoul Wols <raoulwols@gmail.com>