Be a little less eager to notify willSave and didSave
The textDocument/willSave was sent too eagerly.
We should notify only when "save" is set to True in the server's
capabilities. So willSave will never be sent for old-style text doc
sync capabilities.

The textDocument/didSave was also sent too eagerly.
It should only be sent when:
- "save" is a dict, whether empty or not, or
- "save" is a boolean with value True.
So didSave will also never be sent for old-style text doc sync
capabilities.