Remove unused argument to CodeActionsManager.request
It looks the intention of `request` function was to set the value
of passed-in `diagnostics_by_config` but that of course doesn't work as
arguments are not passed by reference in python.

The one place that actually needs diagnostics (modified in this PR)
gets them itself before calling `request` anyway.