Fix buffer modifications while completing
The buffer can be modified by the user, e.g. by typing more characters
while the AC widget is visible. This invalidates all text edits returned
by the language server. We now account for this by keeping an eye on the
buffer modifications via on_text_changed and adjusting the text edits
accordingly.

Another problem was that we called view.change_id() too early. It should be
called just before running the completion command. This accounts for the
prefix removal that ST does.