Extend test server with $test/sendNotification request handler
Allows to trigger any notification from tests.

For example if we want the server to send a 'textDocument/publishDiagnostics'
notification to populate view's diagnostics we'd do:

yield from self.await_client_notification('textDocument/publishDiagnostics', {'params': {}})

On the server side this is a request handler and not a notification handler
as I wanted to be able to await the result so that I would have guarantee
that the notification has been handled when "await_client_notification" call
returned.