reintroduce text_wrap
it removes the vertical scrollbar in most cases.

minihtml accepts a new arg `wrap_width`. When provided it will wrap only 
content that is parsed with the text2html function. 
It won't affect parsing with md2html.

I tried to extract the text wrap logic from the text2html function to
call it just before populating the content in the show_popup.

But I concuded that I need to wrap the text before parsing the links. 
It can happen that the text gets wrapped after <a</b>href="...
which than breaks the links.