Revert back logic for session_for_view.
@rchl can you take a look at this.

If you want.
Checkout the commit before this one.
I have only one session in a file (and that is LSP-typescript, I removed LSP-elint to test this so I am sure I have only one session)

Open a typescript file.
Trigger the AC.
Click the `More` link.

Clicking it should open the Docs popup.
But it doesn't.

Reverting the code to a state like it was before, makes the Docs popup show up