Road to selectors (#931)
* Remove syntaxes, replace with document_selector
* Remove scopes, replace with feature_selector
* Deduce document_selector from languageId if possible
* Deduce feature_selector from document_selector
* Use 'debug' to print the deprecation warning about old "syntaxes" key
* Improve lookup performance of sessions

The document_selector is used to decide when to start a language server.

The feature_selector is used in deciding which language server is "best" for a
point in the document. For example, vscode-html would use feature_selector
to only provide hover/completion capabilities in embedded HTML in PHP.

This can be accomplished with the following feature_selector:

    text.html

Because:

>>> sublime.score_selector("embedding.php text.html", "text.html")
128

Whereas for instance

>>> sublime.score_selector("embedding.php variable.function", "text.html")
0

Meanwhile a PHP language server would have embedding.php as feature_selector
(the feature_selector defaults to the document_selector). But,

>>> sublime.score_selector("embedding.php text.html", "embedding.php")
16

This means a HTML language server with feature_selector text.html would win
over a PHP language server in embedded HTML in PHP.