Introduce the SessionView class
It's an object whose lifetime is controlled by a ViewEventListener
It's an object per view per session. So it's essentially view-state
per session.
Each Session object holds a WeakSet to all of its SessionViews.
Each SessionView holds a strong reference to its Session.
The WindowManager holds weak references to Sessions.
The WindowManager also holds weak references to the ViewEventListeners.
This means the SessionView is the only object that holds a strong
reference to a Session.
Which means that if all views applicable to a Session are removed,
then so is the associated Session.