Register top-level capabilities for each associated view of the Session
Top-level capabilities like definitionProvider and hoverProvider are now
stashed in the view settings, for each view connected to the Session. This will
allow very precise context keys to be made in the keybindings.

It also fixes the problem of ST showing its own hover popup. We disable that
now if and only if there's a session with a hoverProvider connected to the view.

Previously, we disabled it even if the session was not a hoverProvider.