Fix is_subpath_of
The is_subpath_of function does not work for symlinked files. It also
functions incorrectly for case-sensitive filesystems like ext4 on Linux.

For py33, the os.path.realpath function is os.path.abspath for Windows.
So the situation is still not entirely correct for Windows users that
have symlinked files, but we'll  get to that later.

Signed-off-by: Raoul Wols <raoulwols@gmail.com>