Introduce the SessionView class (#1069)
It's an object whose lifetime is controlled by a ViewEventListener
It's an object per view per session. So it's essentially view-state
per session.
Each Session object holds a WeakSet to all of its SessionViews.
Each SessionView holds a strong reference to its Session.
The WindowManager holds weak references to Sessions.
The WindowManager also holds weak references to the ViewEventListeners.
This means the SessionView is the only object that holds a strong
reference to a Session.
Which means that if all views applicable to a Session are removed,
then so is the associated Session (after three seconds).

* Don't send didClose when the session is shutting down
* Turn _sessions into a simple WeakSet, and put exiting in the Session class
* Add can_start to Session, and more tweaks for the DocumentSyncListener
* Register top-level capabilities for each associated view of the Session

Top-level capabilities like definitionProvider and hoverProvider are now
stashed in the view settings, for each view connected to the Session. This will
allow very precise context keys to be made in the keybindings.

It also fixes the problem of ST showing its own hover popup. We disable that
now if and only if there's a session with a hoverProvider connected to the view.

Previously, we disabled it even if the session was not a hoverProvider.

* Keep dequeing pending listeners when for the current listener
  there's no new config
* Do session administration on ST's worker thread
* Even unregister session views on the async thread
* Register new sessions on the async thread
* Optimize score_document with an LRU cache, it's very slow during startup
* Fix is_subpath_of

The is_subpath_of function does not work for symlinked files. It also
functions incorrectly for case-sensitive filesystems like ext4 on Linux.

For py33, the os.path.realpath function is os.path.abspath for Windows.
So the situation is still not entirely correct for Windows users that
have symlinked files, but we'll  get to that later.


* Check for clones
* Replace __contains__ by contains in ProjectFolders class
* Reorder methods in WindowManager for better readability
* Remove unused DocumentsState class
* Bail out early in ProjectFolders.contains if the view is dead
* Move debounced end_async to the Session class
* Rename match_document -> match_scope