Do not stash capabilities in the view settings
Instead query whether there exists a session with the capability
in the on_query_context callback from sublime text.

We do still need to listen for the hoverProvider capability, because we need to
disable ST's built-in hover functionality once there exists at least one
hoverProvider (and re-enable it again once all hoverProviders are gone).

These changes allow us to check for

- textDocumentSync.willSave
- codeActionProvider.codeActionKinds

in the keybindings.