Fix broken LspResolveDocs run command
I guess everybody just uses F12. When you click on "More" with the mouse, the
run command will receive an extra "event" parameter. This is because the base
class LspTextCommand wants events. We don't actually do anything with that
event position in the run method so we can ignore it.