Remove the session views lock (#1169)
When you run everything on a single thread, there is no need for locks. It's a
bit tricky because we have to use on_pre_save. There's no guarantee that
on_pre_save_async is invoked *before* the view is saved. And making sure
that purge_changes_async is called from the async thread is also somewhat
tricky. Things that might be affected:

- completion handling calling purge_changes_async
- signature help calling purge_changes_async
- save command calling purge_changes_async
- Switch SaveTask to run on ST's async thread

To sync with purge_changes_async that now runs on STs async thread,
SaveTask also needs to run there.

As part of that, the CodeActionsCollector was changed to call response
handler on ST's async thread (previously called it on the main thread).

The changes are not super through in terms of naming things. I haven't
changed all methods of SaveTask to have "_async` suffix since whole
task should run async. Only changed "run" to run_async" to make it
clear on the entry point.

Also, since responses to rpc calls are triggered on custom async thread,
I'm switching over to STs async thread in those cases. (Maybe the RPC
itself should call responses on either ST's main or async thread, for
consistency?).

- Fold textDocument/willSave notification into lsp_save command
- Do not stash capabilities in the view settings

Instead query whether there exists a session with the capability
in the on_query_context callback from sublime text.

We do still need to listen for the hoverProvider capability, because we need to
disable ST's built-in hover functionality once there exists at least one
hoverProvider (and re-enable it again once all hoverProviders are gone).

These changes allow us to check for

- textDocumentSync.willSave
- codeActionProvider.codeActionKinds

in the keybindings.

- Add "textDocumentSync.willSaveWaitUntil" to the capabilities of the keybindings
- Fix textDocument/didChange triggering at wrong time

Co-authored-by: Rafał Chłodnicki <rchl2k@gmail.com>