allow sending highlight request when there are multiple cursors
This is mostly done when someone uses ctrl+d multiple times
and changes the text.

Before the highlight request wont be send and we wont see the underlines.
With this commit we send the highlight request and we will see the underlines.

The downside is the a user can place cursors on different words that
are not the same symbol, but we that is just an edge case
(that is not worth adding).