Capabilities per SessionBuffer (#1282)
Each SessionBuffer has its own capabilities dictionary. This is
to account for dynamic registrations which use a DocumentSelector.

This diff makes OmniSharp somewhat work OK. It is a good example
of a server that registers *all* of its capabilities dynamically,
even textDocumentSync. Because textDocumentSync can be registered
dynamically, we need to account for the possibility of a
SessionBuffer's life starting out without document synchronization,
only to later "obtain" document sync dynamically. This is accounted
for in this diff.

A follow-up pull request will implement the `pattern` requirement
of a DocumentFilter.

This diff also implements the notion of a DocumentSelector. It is
a straightforward translation from the specification. A language server
may register a capability with a DocumentSelector consisting of a
`language` entry. This entry is the "Language ID". Since an arbitrary
language server doesn't know about SublimeText base scopes, we will
have to maintain this translation in a global settings file via
language-ids.sublime-settings.

We may have to swap the LanguageConfig class with the DocumentSelector.