Fix diagnostics not being displayed in the view for virtual files
Diagnostics were not displays for default preferences, for example,
since those files don't physically exist on disk.

The purpose of the original "is_same_file()" check was to match also files
that differed in case of the drive letter but if the file name matches
exactly then we don't have to call it.