Print a debug message instead of asserting when registering a plugin
When a plugin was already registered, and then called `register_plugin`
again, we would hit an assertion error. Let's print a debug message
instead so that we can make a smooth transition to explicitly registering
plugins instead of the current way where we register them "immediately"
in _forcefully_register_plugins.